refactor(tail): crane-mcp tools+lib batch — 12 files cleaned
Tail-batch refactors across crane-mcp lib + tools. Multiple files brought
to 0 lint errors via helper extraction and structural splits.

- lib/crane-api.ts: rewritten as thin barrel + CraneApi extends
  CraneApiSchedule. New: crane-api-base.ts (CraneApiBase core),
  crane-api-schedule.ts (CraneApiSchedule extends Base), crane-api-types.ts
  (shared domain types), crane-api-extended-types.ts (telemetry +
  notification + fleet types with Zod schemas).
- lib/token-tracker.ts: added missing readFileSync import.
- lib/pr-merge-gate.ts, lib/session-log.ts: minor structural fixes.
- tools/doc-audit.ts: bundled 6-param functions into FixContext object.
- tools/handoff.ts: extracted resolveVenture, buildAndSubmitHandoff,
  postSessionActivityEvents.
- tools/docs-drift-audit.ts: split into schema/executor/formatter +
  barrel re-exports. New: docs-drift-checks.ts (six check functions),
  docs-drift-utils.ts (filesystem/parsing).
- tools/{deploy-heartbeat,notifications,preflight,schedule}.ts: helper
  extractions for complexity/length compliance.
- cli/ssh-auth.ts: minor fix.

Public API preserved. 764/764 tests pass.
