+++ packages/crane-mcp/src/tools/schedule.ts
@@ -170,6 +170,277 @@ function getApiKey(): string | null {
   return process.env.CRANE_CONTEXT_KEY || null
 }
 
+async function scheduleList(api: CraneApi, input: ScheduleInput): Promise<ScheduleResult> {
+  try {
+    const briefing = await api.getScheduleBriefing(input.scope)
+    let message = '## Schedule Briefing\n\n'
+    message += formatBriefingTable(briefing.items)
+    if (briefing.items.length > 0) {
+      const parts: string[] = []
+      if (briefing.overdue_count > 0) parts.push(`${briefing.overdue_count} overdue`)
+      if (briefing.due_count > 0) parts.push(`${briefing.due_count} due`)
+      if (briefing.untracked_count > 0) parts.push(`${briefing.untracked_count} untracked`)
+      message += `\n${parts.join(', ')}`
+    }
+    return { success: true, message }
+  } catch (error) {
+    return {
+      success: false,
+      message: `Failed to fetch schedule briefing: ${error instanceof Error ? error.message : 'Unknown error'}`,
+    }
+  }
+}
+
+async function scheduleItems(api: CraneApi): Promise<ScheduleResult> {
+  try {
+    const response = await api.getScheduleItems()
+    let message = '## All Schedule Items\n\n'
+    message += formatItemsTable(response.items)
+    message += `\n${response.count} items total`
+    return { success: true, message }
+  } catch (error) {
+    return {
+      success: false,
+      message: `Failed to fetch schedule items: ${error instanceof Error ? error.message : 'Unknown error'}`,
+    }
+  }
+}
+
+async function scheduleLinkCalendar(api: CraneApi, input: ScheduleInput): Promise<ScheduleResult> {
+  if (!input.name) {
+    return {
+      success: false,
+      message: 'Schedule item name is required for the link-calendar action.',
+    }
+  }
+  try {
+    const gcalId = input.gcal_event_id === undefined ? null : input.gcal_event_id
+    const response = await api.linkScheduleCalendar(input.name, gcalId)
+    const action = response.gcal_event_id ? 'Linked' : 'Unlinked'
+    return {
+      success: true,
+      message: `${action} calendar for **${response.name}** (gcal_event_id: ${response.gcal_event_id || 'null'})`,
+    }
+  } catch (error) {
+    return {
+      success: false,
+      message: `Failed to link calendar: ${error instanceof Error ? error.message : 'Unknown error'}`,
+    }
+  }
+}
+
+async function scheduleComplete(api: CraneApi, input: ScheduleInput): Promise<ScheduleResult> {
+  if (!input.name) {
+    return { success: false, message: 'Schedule item name is required for the complete action.' }
+  }
+  if (!input.result) {
+    return {
+      success: false,
+      message: 'Result is required for the complete action (success/warning/failure/skipped).',
+    }
+  }
+  try {
+    const response = await api.completeScheduleItem(input.name, {
+      result: input.result,
+      summary: input.summary,
+      completed_by: input.completed_by,
+    })
+    let message = `Completed **${input.name}** at ${response.completed_at} (result: ${response.result}). Next due: ${response.next_due_date || 'unknown'}.`
+    message += '\n\nCalendar sync:'
+    if (response.gcal_event_id) {
+      message += `\n1. Update Google Calendar: gcal_update_event(event_id: "${response.gcal_event_id}", date: ${response.next_due_date})`
+    } else {
+      message += '\n1. No Google Calendar event linked - consider creating one'
+    }
+    message += `\n2. Mark Apple Reminder complete (if exists): search for reminder matching "${input.name}"`
+    return { success: true, message }
+  } catch (error) {
+    return {
+      success: false,
+      message: `Failed to complete schedule item: ${error instanceof Error ? error.message : 'Unknown error'}`,
+    }
+  }
+}
+
+async function schedulePlannedEvents(api: CraneApi, input: ScheduleInput): Promise<ScheduleResult> {
+  if (!input.from || !input.to) {
+    return {
+      success: false,
+      message: 'Both from and to dates are required for the planned-events action.',
+    }
+  }
+  try {
+    const events = await api.getPlannedEvents(input.from, input.to, input.type)
+    if (events.length === 0) {
+      return { success: true, message: 'No planned events found in the specified range.' }
+    }
+    let table = '| ID | Date | Venture | Title | Type | Status | GCal |\n'
+    table += '|----|------|---------|-------|------|--------|------|\n'
+    for (const e of events) {
+      const gcal = e.gcal_event_id ? 'linked' : '-'
+      table += `| ${e.id} | ${e.event_date} | ${e.venture.toUpperCase()} | ${e.title} | ${e.type} | ${e.sync_status} | ${gcal} |\n`
+    }
+    return { success: true, message: `## Planned Events\n\n${table}\n${events.length} events` }
+  } catch (error) {
+    return {
+      success: false,
+      message: `Failed to fetch planned events: ${error instanceof Error ? error.message : 'Unknown error'}`,
+    }
+  }
+}
+
+async function schedulePlannedEventCreate(
+  api: CraneApi,
+  input: ScheduleInput
+): Promise<ScheduleResult> {
+  if (!input.event_date || !input.venture || !input.title || !input.start_time || !input.end_time) {
+    return {
+      success: false,
+      message:
+        'event_date, venture, title, start_time, and end_time are required for planned-event-create.',
+    }
+  }
+  try {
+    const eventType = input.type as 'planned' | 'actual' | 'cancelled' | undefined
+    const event = await api.createPlannedEvent({
+      event_date: input.event_date,
+      venture: input.venture,
+      title: input.title,
+      start_time: input.start_time,
+      end_time: input.end_time,
+      gcal_event_id: input.gcal_event_id,
+      type: eventType,
+    })
+    return {
+      success: true,
+      message: `Created planned event **${event.title}** on ${event.event_date} (${event.id})`,
+    }
+  } catch (error) {
+    return {
+      success: false,
+      message: `Failed to create planned event: ${error instanceof Error ? error.message : 'Unknown error'}`,
+    }
+  }
+}
+
+async function schedulePlannedEventUpdate(
+  api: CraneApi,
+  input: ScheduleInput
+): Promise<ScheduleResult> {
+  if (!input.id) {
+    return { success: false, message: 'Event id is required for planned-event-update.' }
+  }
+  try {
+    const updates: Partial<
+      Pick<
+        import('../lib/crane-api.js').PlannedEvent,
+        'type' | 'start_time' | 'end_time' | 'sync_status' | 'gcal_event_id'
+      >
+    > = {}
+    if (input.type !== undefined) updates.type = input.type as 'planned' | 'actual' | 'cancelled'
+    if (input.start_time !== undefined) updates.start_time = input.start_time
+    if (input.end_time !== undefined) updates.end_time = input.end_time
+    if (input.sync_status !== undefined) updates.sync_status = input.sync_status
+    if (input.gcal_event_id !== undefined) updates.gcal_event_id = input.gcal_event_id
+    const event = await api.updatePlannedEvent(input.id, updates)
+    return {
+      success: true,
+      message: `Updated planned event **${event.title}** (${event.id}): type=${event.type}, sync_status=${event.sync_status}`,
+    }
+  } catch (error) {
+    return {
+      success: false,
+      message: `Failed to update planned event: ${error instanceof Error ? error.message : 'Unknown error'}`,
+    }
+  }
+}
+
+async function schedulePlannedEventsClear(
+  api: CraneApi,
+  input: ScheduleInput
+): Promise<ScheduleResult> {
+  if (!input.from) {
+    return { success: false, message: 'from date is required for planned-events-clear.' }
+  }
+  try {
+    const result = await api.clearPlannedEvents(input.from)
+    return {
+      success: true,
+      message: `Cleared ${result.deleted} planned events from ${input.from} onwards.`,
+    }
+  } catch (error) {
+    return {
+      success: false,
+      message: `Failed to clear planned events: ${error instanceof Error ? error.message : 'Unknown error'}`,
+    }
+  }
+}
+
+function formatSessionHistoryBlock(
+  entry: import('../lib/crane-api.js').SessionHistoryEntry,
+  block: import('../lib/crane-api.js').SessionHistoryEntry['blocks'][number]
+): string {
+  const startDate = new Date(block.start)
+  const endDate = new Date(block.end)
+  const timeOpts: Intl.DateTimeFormatOptions = {
+    hour: 'numeric',
+    minute: '2-digit',
+    timeZone: 'America/Phoenix',
+  }
+  const start = startDate.toLocaleTimeString('en-US', timeOpts)
+  const startDay = new Date(
+    startDate.toLocaleString('en-US', { timeZone: 'America/Phoenix' })
+  ).getDate()
+  const endDay = new Date(
+    endDate.toLocaleString('en-US', { timeZone: 'America/Phoenix' })
+  ).getDate()
+  const end =
+    endDay !== startDay
+      ? endDate.toLocaleString('en-US', {
+          ...timeOpts,
+          month: 'numeric',
+          day: 'numeric',
+          timeZone: 'America/Phoenix',
+        })
+      : endDate.toLocaleTimeString('en-US', timeOpts)
+  const detail =
+    (block.hosts?.[0] || '-') +
+    (block.repos?.[0] ? ' · ' + block.repos[0].split('/').pop() : '') +
+    (block.issues?.[0] ? ' #' + String(block.issues[0]) : '')
+  return `| ${entry.work_date} | ${entry.venture.toUpperCase()} | ${start} | ${end} | ${block.session_count} | ${detail} |\n`
+}
+
+async function scheduleSessionHistory(
+  api: CraneApi,
+  input: ScheduleInput
+): Promise<ScheduleResult> {
+  try {
+    const days = input.days || 7
+    const entries = await api.getSessionHistory(days)
+    if (entries.length === 0) {
+      return { success: true, message: `No ended sessions found in the last ${days} days.` }
+    }
+    let table = '| Date | Venture | Start | End | Sessions | Detail |\n'
+    table += '|------|---------|-------|-----|----------|--------|\n'
+    let totalBlocks = 0
+    for (const entry of entries) {
+      for (const block of entry.blocks) {
+        totalBlocks++
+        table += formatSessionHistoryBlock(entry, block)
+      }
+    }
+    return {
+      success: true,
+      message: `## Session History (${days} days)\n\n${table}\n${totalBlocks} blocks across ${entries.length} venture-days`,
+    }
+  } catch (error) {
+    return {
+      success: false,
+      message: `Failed to fetch session history: ${error instanceof Error ? error.message : 'Unknown error'}`,
+    }
+  }
+}
+
 export async function executeSchedule(input: ScheduleInput): Promise<ScheduleResult> {
   const apiKey = getApiKey()
   if (!apiKey) {
@@ -181,281 +452,29 @@ export async function executeSchedule(input: ScheduleInput): Promise<ScheduleRes
 
   const api = new CraneApi(apiKey, getApiBase())
 
-  if (input.action === 'list') {
-    try {
-      const briefing = await api.getScheduleBriefing(input.scope)
-
-      let message = '## Schedule Briefing\n\n'
-      message += formatBriefingTable(briefing.items)
-
-      if (briefing.items.length > 0) {
-        const parts: string[] = []
-        if (briefing.overdue_count > 0) parts.push(`${briefing.overdue_count} overdue`)
-        if (briefing.due_count > 0) parts.push(`${briefing.due_count} due`)
-        if (briefing.untracked_count > 0) parts.push(`${briefing.untracked_count} untracked`)
-        message += `\n${parts.join(', ')}`
-      }
-
-      return { success: true, message }
-    } catch (error) {
+  switch (input.action) {
+    case 'list':
+      return scheduleList(api, input)
+    case 'items':
+      return scheduleItems(api)
+    case 'link-calendar':
+      return scheduleLinkCalendar(api, input)
+    case 'complete':
+      return scheduleComplete(api, input)
+    case 'planned-events':
+      return schedulePlannedEvents(api, input)
+    case 'planned-event-create':
+      return schedulePlannedEventCreate(api, input)
+    case 'planned-event-update':
+      return schedulePlannedEventUpdate(api, input)
+    case 'planned-events-clear':
+      return schedulePlannedEventsClear(api, input)
+    case 'session-history':
+      return scheduleSessionHistory(api, input)
+    default:
       return {
         success: false,
-        message: `Failed to fetch schedule briefing: ${error instanceof Error ? error.message : 'Unknown error'}`,
+        message: `Unknown action: ${String(input.action)}. Use "list", "complete", "items", "link-calendar", "planned-events", "planned-event-create", "planned-event-update", "planned-events-clear", or "session-history".`,
       }
-    }
-  }
-
-  if (input.action === 'items') {
-    try {
-      const response = await api.getScheduleItems()
-
-      let message = '## All Schedule Items\n\n'
-      message += formatItemsTable(response.items)
-      message += `\n${response.count} items total`
-
-      return { success: true, message }
-    } catch (error) {
-      return {
-        success: false,
-        message: `Failed to fetch schedule items: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      }
-    }
-  }
-
-  if (input.action === 'link-calendar') {
-    if (!input.name) {
-      return {
-        success: false,
-        message: 'Schedule item name is required for the link-calendar action.',
-      }
-    }
-
-    try {
-      const gcalId = input.gcal_event_id === undefined ? null : input.gcal_event_id
-      const response = await api.linkScheduleCalendar(input.name, gcalId)
-
-      const action = response.gcal_event_id ? 'Linked' : 'Unlinked'
-      return {
-        success: true,
-        message: `${action} calendar for **${response.name}** (gcal_event_id: ${response.gcal_event_id || 'null'})`,
-      }
-    } catch (error) {
-      return {
-        success: false,
-        message: `Failed to link calendar: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      }
-    }
-  }
-
-  if (input.action === 'complete') {
-    if (!input.name) {
-      return {
-        success: false,
-        message: 'Schedule item name is required for the complete action.',
-      }
-    }
-    if (!input.result) {
-      return {
-        success: false,
-        message: 'Result is required for the complete action (success/warning/failure/skipped).',
-      }
-    }
-
-    try {
-      const response = await api.completeScheduleItem(input.name, {
-        result: input.result,
-        summary: input.summary,
-        completed_by: input.completed_by,
-      })
-
-      let message = `Completed **${input.name}** at ${response.completed_at} (result: ${response.result}). Next due: ${response.next_due_date || 'unknown'}.`
-
-      // Add calendar sync orchestration hints
-      message += '\n\nCalendar sync:'
-      if (response.gcal_event_id) {
-        message += `\n1. Update Google Calendar: gcal_update_event(event_id: "${response.gcal_event_id}", date: ${response.next_due_date})`
-      } else {
-        message += '\n1. No Google Calendar event linked - consider creating one'
-      }
-      message += `\n2. Mark Apple Reminder complete (if exists): search for reminder matching "${input.name}"`
-
-      return {
-        success: true,
-        message,
-      }
-    } catch (error) {
-      return {
-        success: false,
-        message: `Failed to complete schedule item: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      }
-    }
-  }
-
-  if (input.action === 'planned-events') {
-    if (!input.from || !input.to) {
-      return {
-        success: false,
-        message: 'Both from and to dates are required for the planned-events action.',
-      }
-    }
-    try {
-      const events = await api.getPlannedEvents(input.from, input.to, input.type)
-      if (events.length === 0) {
-        return { success: true, message: 'No planned events found in the specified range.' }
-      }
-      let table = '| ID | Date | Venture | Title | Type | Status | GCal |\n'
-      table += '|----|------|---------|-------|------|--------|------|\n'
-      for (const e of events) {
-        const gcal = e.gcal_event_id ? 'linked' : '-'
-        table += `| ${e.id} | ${e.event_date} | ${e.venture.toUpperCase()} | ${e.title} | ${e.type} | ${e.sync_status} | ${gcal} |\n`
-      }
-      return { success: true, message: `## Planned Events\n\n${table}\n${events.length} events` }
-    } catch (error) {
-      return {
-        success: false,
-        message: `Failed to fetch planned events: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      }
-    }
-  }
-
-  if (input.action === 'planned-event-create') {
-    if (
-      !input.event_date ||
-      !input.venture ||
-      !input.title ||
-      !input.start_time ||
-      !input.end_time
-    ) {
-      return {
-        success: false,
-        message:
-          'event_date, venture, title, start_time, and end_time are required for planned-event-create.',
-      }
-    }
-    try {
-      const event = await api.createPlannedEvent({
-        event_date: input.event_date,
-        venture: input.venture,
-        title: input.title,
-        start_time: input.start_time,
-        end_time: input.end_time,
-        gcal_event_id: input.gcal_event_id,
-        type: input.type as 'planned' | 'actual' | 'cancelled' | undefined,
-      })
-      return {
-        success: true,
-        message: `Created planned event **${event.title}** on ${event.event_date} (${event.id})`,
-      }
-    } catch (error) {
-      return {
-        success: false,
-        message: `Failed to create planned event: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      }
-    }
-  }
-
-  if (input.action === 'planned-event-update') {
-    if (!input.id) {
-      return { success: false, message: 'Event id is required for planned-event-update.' }
-    }
-    try {
-      const updates: Record<string, unknown> = {}
-      if (input.type !== undefined) updates.type = input.type
-      if (input.start_time !== undefined) updates.start_time = input.start_time
-      if (input.end_time !== undefined) updates.end_time = input.end_time
-      if (input.sync_status !== undefined) updates.sync_status = input.sync_status
-      if (input.gcal_event_id !== undefined) updates.gcal_event_id = input.gcal_event_id
-      const event = await api.updatePlannedEvent(input.id, updates as any)
-      return {
-        success: true,
-        message: `Updated planned event **${event.title}** (${event.id}): type=${event.type}, sync_status=${event.sync_status}`,
-      }
-    } catch (error) {
-      return {
-        success: false,
-        message: `Failed to update planned event: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      }
-    }
-  }
-
-  if (input.action === 'planned-events-clear') {
-    if (!input.from) {
-      return { success: false, message: 'from date is required for planned-events-clear.' }
-    }
-    try {
-      const result = await api.clearPlannedEvents(input.from)
-      return {
-        success: true,
-        message: `Cleared ${result.deleted} planned events from ${input.from} onwards.`,
-      }
-    } catch (error) {
-      return {
-        success: false,
-        message: `Failed to clear planned events: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      }
-    }
-  }
-
-  if (input.action === 'session-history') {
-    try {
-      const days = input.days || 7
-      const entries = await api.getSessionHistory(days)
-      if (entries.length === 0) {
-        return { success: true, message: `No ended sessions found in the last ${days} days.` }
-      }
-      let table = '| Date | Venture | Start | End | Sessions | Detail |\n'
-      table += '|------|---------|-------|-----|----------|--------|\n'
-      let totalBlocks = 0
-      for (const entry of entries) {
-        for (const block of entry.blocks) {
-          totalBlocks++
-          const startDate = new Date(block.start)
-          const endDate = new Date(block.end)
-          const timeOpts: Intl.DateTimeFormatOptions = {
-            hour: 'numeric',
-            minute: '2-digit',
-            timeZone: 'America/Phoenix',
-          }
-          const start = startDate.toLocaleTimeString('en-US', timeOpts)
-
-          // Show date on end time when the block spans midnight
-          const startDay = new Date(
-            startDate.toLocaleString('en-US', { timeZone: 'America/Phoenix' })
-          ).getDate()
-          const endDay = new Date(
-            endDate.toLocaleString('en-US', { timeZone: 'America/Phoenix' })
-          ).getDate()
-          const end =
-            endDay !== startDay
-              ? endDate.toLocaleString('en-US', {
-                  ...timeOpts,
-                  month: 'numeric',
-                  day: 'numeric',
-                  timeZone: 'America/Phoenix',
-                })
-              : endDate.toLocaleTimeString('en-US', timeOpts)
-
-          const detail =
-            (block.hosts?.[0] || '-') +
-            (block.repos?.[0] ? ' · ' + block.repos[0].split('/').pop() : '') +
-            (block.issues?.[0] ? ' #' + block.issues[0] : '')
-          table += `| ${entry.work_date} | ${entry.venture.toUpperCase()} | ${start} | ${end} | ${block.session_count} | ${detail} |\n`
-        }
-      }
-      return {
-        success: true,
-        message: `## Session History (${days} days)\n\n${table}\n${totalBlocks} blocks across ${entries.length} venture-days`,
-      }
-    } catch (error) {
-      return {
-        success: false,
-        message: `Failed to fetch session history: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      }
-    }
-  }
-
-  return {
-    success: false,
-    message: `Unknown action: ${input.action}. Use "list", "complete", "items", "link-calendar", "planned-events", "planned-event-create", "planned-event-update", "planned-events-clear", or "session-history".`,
   }
 }
